@@ -5,19 +5,17 @@ publish_to: "none"
 version: 1.0.3+0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.0
-  flutter_launcher_icons: ^0.8.1
   file_picker: ^3.0.0
-  webview_flutter: ^2.0.2
-  responsive_flutter: ^0.0.4
-  metadata_fetch: ^0.3.4
-  shimmer: ^1.1.2
   url_launcher: ^6.0.3
+  shimmer: ^3.0.0
+  metadata_fetch: ^0.4.1
+  webview_flutter: ^4.2.2
 
 dev_dependencies:
   flutter_test:
