@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:responsive_flutter/responsive_flutter.dart';
 
 class HomeCard extends StatelessWidget {
   final AssetImage backgroundImage;
@@ -7,10 +6,10 @@ class HomeCard extends StatelessWidget {
   final VoidCallback onTap;
 
   HomeCard(
-      {Key key,
-      @required this.backgroundImage,
-      @required this.title,
-      @required this.onTap})
+      {Key? key,
+      required this.backgroundImage,
+      required this.title,
+      required this.onTap})
       : super(key: key);
 
   @override
@@ -26,19 +25,19 @@ class HomeCard extends StatelessWidget {
           children: <Widget>[
             Container(
               width: MediaQuery.of(context).size.width,
-              height: ResponsiveFlutter.of(context).verticalScale(100.0),
+              // height: ResponsiveFlutter.of(context).verticalScale(100.0),
               child: Image(
                 image: backgroundImage,
                 fit: BoxFit.fill,
               ),
             ),
             Container(
-              padding: EdgeInsets.only(
-                left: ResponsiveFlutter.of(context).moderateScale(10.0),
-                top: ResponsiveFlutter.of(context).moderateScale(12.0),
-                right: ResponsiveFlutter.of(context).moderateScale(10.0),
-                bottom: ResponsiveFlutter.of(context).moderateScale(0.0),
-              ),
+              // padding: EdgeInsets.only(
+              //   left: ResponsiveFlutter.of(context).moderateScale(10.0),
+              //   top: ResponsiveFlutter.of(context).moderateScale(12.0),
+              //   right: ResponsiveFlutter.of(context).moderateScale(10.0),
+              //   bottom: ResponsiveFlutter.of(context).moderateScale(0.0),
+              // ),
               width: MediaQuery.of(context).size.width,
               child: title,
             ),
