@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:agried/models.dart';
-import 'package:responsive_flutter/responsive_flutter.dart';
 
 class StudentProfileCard extends StatelessWidget {
   StudentProfileCard({
-    Key key,
-    @required this.profileImage,
-    @required this.student,
-    @required this.onTap,
+    Key? key,
+    required this.profileImage,
+    required this.student,
+    required this.onTap,
   }) : super(key: key);
 
   // Change this to NetworkImage once the API is done.
@@ -18,10 +17,10 @@ class StudentProfileCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     CircleAvatar studentAvatar = CircleAvatar(
-      radius: ResponsiveFlutter.of(context).scale(55),
+      // radius: ResponsiveFlutter.of(context).scale(55),
       backgroundColor: Colors.teal[400],
       child: CircleAvatar(
-        radius: ResponsiveFlutter.of(context).scale(50),
+        // radius: ResponsiveFlutter.of(context).scale(50),
         backgroundImage: profileImage,
       ),
     );
@@ -32,7 +31,7 @@ class StudentProfileCard extends StatelessWidget {
         Text(
           student.name,
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
             fontWeight: FontWeight.bold,
             color: Colors.blue[900],
           ),
@@ -40,7 +39,7 @@ class StudentProfileCard extends StatelessWidget {
         Text(
           'Kelas ${student.className}',
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
             fontWeight: FontWeight.bold,
             color: Colors.blue[900],
           ),
@@ -48,7 +47,7 @@ class StudentProfileCard extends StatelessWidget {
         Text(
           student.major,
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
             fontWeight: FontWeight.bold,
             color: Colors.blue[900],
           ),
@@ -56,7 +55,7 @@ class StudentProfileCard extends StatelessWidget {
         Text(
           student.school.name,
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
             fontWeight: FontWeight.bold,
             color: Colors.blue[900],
           ),
@@ -64,7 +63,7 @@ class StudentProfileCard extends StatelessWidget {
         Text(
           'No: ${student.serialNo}',
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
             fontWeight: FontWeight.bold,
             color: Colors.blue[900],
           ),
@@ -83,7 +82,7 @@ class StudentProfileCard extends StatelessWidget {
             gradient: new LinearGradient(
               colors: [
                 Colors.white,
-                Colors.orange[700],
+                Colors.orange,
               ],
               begin: const FractionalOffset(0.1, 0.0),
               end: const FractionalOffset(1.2, 0.0),
@@ -101,38 +100,37 @@ class StudentProfileCard extends StatelessWidget {
               ),
               Padding(
                 padding: EdgeInsets.all(
-                  ResponsiveFlutter.of(context).moderateScale(1.0),
+                  // ResponsiveFlutter.of(context).moderateScale(1.0),
+                  1.0,
                 ),
                 child: Row(
                   children: <Widget>[
                     studentAvatar,
                     SizedBox(
-                      width: ResponsiveFlutter.of(context).scale(12.0),
+                      // width: ResponsiveFlutter.of(context).scale(12.0),
                     ),
                     studentDetails,
                     SizedBox(
-                      width: ResponsiveFlutter.of(context).scale(20.0),
+                      // width: ResponsiveFlutter.of(context).scale(20.0),
                     ),
                     Column(
                       mainAxisAlignment: MainAxisAlignment.center,
                       crossAxisAlignment: CrossAxisAlignment.center,
                       children: <Widget>[
                         Image(
-                          height:
-                              ResponsiveFlutter.of(context).verticalScale(55.0),
+                          // height: ResponsiveFlutter.of(context).verticalScale(55.0),
                           image: AssetImage(
                             'assets/icons/clock.png',
                           ),
                         ),
                         SizedBox(
-                          height:
-                              ResponsiveFlutter.of(context).verticalScale(1.0),
+                          // height: ResponsiveFlutter.of(context).verticalScale(1.0),
                         ),
                         Text(
                           'Next class',
                           style: TextStyle(
-                            fontSize:
-                                ResponsiveFlutter.of(context).fontSize(1.5),
+                            // fontSize:
+                            //     ResponsiveFlutter.of(context).fontSize(1.5),
                             fontWeight: FontWeight.bold,
                             color: Colors.blue[900],
                           ),
@@ -140,8 +138,7 @@ class StudentProfileCard extends StatelessWidget {
                         Text(
                           'in 22 mins',
                           style: TextStyle(
-                            fontSize:
-                                ResponsiveFlutter.of(context).fontSize(1.5),
+                            // fontSize: ResponsiveFlutter.of(context).fontSize(1.5),
                             fontWeight: FontWeight.bold,
                             color: Colors.blue[900],
                           ),
