@@ -1,11 +1,9 @@
-import 'package:flutter/material.dart';
-
 class CalculatorUnit {
   final String name;
   final String unit;
 
   CalculatorUnit({
-    @required this.name,
-    @required this.unit,
+    required this.name,
+    required this.unit,
   });
 }
