@@ -1,16 +1,15 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
-import 'package:responsive_flutter/responsive_flutter.dart';
 
 class OnlineNewsScreen extends StatefulWidget {
   final String title;
   final String url;
 
   OnlineNewsScreen({
-    Key key,
-    @required this.title,
-    @required this.url,
+    Key? key,
+    required this.title,
+    required this.url,
   }) : super(key: key);
 
   @override
@@ -23,21 +22,30 @@ class OnlineNewsScreen extends StatefulWidget {
 class _OnlineNewsScreenState extends State<OnlineNewsScreen> {
   final String title;
   final String url;
+  late WebViewController webViewController;
 
-  bool _loading = true;
+  late int _progress;
 
   _OnlineNewsScreenState({
-    @required this.title,
-    @required this.url,
+    required this.title,
+    required this.url,
   });
 
   @override
   void initState() {
     super.initState();
 
-    if (Platform.isAndroid) {
-      WebView.platform = SurfaceAndroidWebView();
-    }
+    webViewController = WebViewController();
+    webViewController.setJavaScriptMode(JavaScriptMode.unrestricted);
+    webViewController.setNavigationDelegate(
+      NavigationDelegate(
+        onProgress: (int progress) {
+          setState(() {
+            _progress = progress;
+          });
+        },
+      ),
+    );
   }
 
   @override
@@ -50,23 +58,21 @@ class _OnlineNewsScreenState extends State<OnlineNewsScreen> {
           overflow: TextOverflow.ellipsis,
           softWrap: true,
           style: TextStyle(
-            fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
+            // fontSize: ResponsiveFlutter.of(context).fontSize(2.0),
           ),
         ),
         bottom: PreferredSize(
           preferredSize: Size.zero,
-          child: _loading ? LinearProgressIndicator() : SizedBox(),
+          child: LinearProgressIndicator(
+            value: _progress / 100,
+            backgroundColor: Colors.white,
+            valueColor: AlwaysStoppedAnimation<Color>(
+              Colors.blue,
+            ),
+          ),
         ),
       ),
-      body: WebView(
-        initialUrl: url,
-        javascriptMode: JavascriptMode.unrestricted,
-        onPageFinished: (String finished) {
-          setState(() {
-            _loading = false;
-          });
-        },
-      ),
+      body: WebViewWidget(controller: webViewController),
     );
   }
 }
