@@ -2,6 +2,6 @@ class School {
   final String name;
 
   School({
-    this.name,
+    required this.name,
   });
 }
